requirements.txt
asgiref==3.9.1
asttokens==3.0.0
beautifulsoup4==4.13.4
black==25.1.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.15
decorator==5.2.1
decouple==0.0.7
dicttoxml==1.7.16
dj-database-url==3.0.1
Django==5.2.5
djangorestframework==3.16.1
dnspython==1.16.0
elementpath==5.0.3
executing==2.2.0
gunicorn==23.0.0
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
lxml==6.0.0
matplotlib-inline==0.1.7
mypy_extensions==1.1.0
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pygame==2.6.1
pygame-ce==2.5.5
Pygments==2.19.2
pymongo==3.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pywin32==311
pyzmq==27.0.0
requests==2.32.4
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
tabulate==0.9.0
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
whitenoise==6.9.0
xmlschema==4.1.0
